CNDB-11434: Make all tests in TrieMemtableIndexTest be run with both offheap and onheap to improve test coverage
